+++ src/main/java/io/github/moulberry/notenoughupdates/NEUApi.java
@@ -3,7 +3,7 @@ package io.github.moulberry.notenoughupdates;
 import net.minecraftforge.fml.relauncher.ReflectionHelper;
 
 public class NEUApi {
-    static boolean disableInventoryButtons = false;
+    public static boolean disableInventoryButtons = false;
 
     public static void setInventoryButtonsToDisabled() {
         disableInventoryButtons = true;
